# Conflicts:
#	pom.xml
#	src/main/java/com/amc/api/Config/CorsConfiguration.java
#	src/main/java/com/amc/api/Config/ModelMapperConfiguration.java
#	src/main/java/com/amc/api/Config/SecurityConfig.java
#	src/main/java/com/amc/api/Controllers/AddressController.java
#	src/main/java/com/amc/api/Controllers/AuthenticationController.java
#	src/main/java/com/amc/api/Controllers/CategoryController.java
#	src/main/java/com/amc/api/Controllers/Order/OrderController.java
#	src/main/java/com/amc/api/Controllers/Product/FileProductController.java
#	src/main/java/com/amc/api/Controllers/Product/ProductController.java
#	src/main/java/com/amc/api/Controllers/UserController.java
#	src/main/java/com/amc/api/Entities/Customer.java
#	src/main/java/com/amc/api/Entities/Product/Product.java
#	src/main/java/com/amc/api/Entities/User.java
#	src/main/java/com/amc/api/Enums/UserRoleEnum.java
#	src/main/java/com/amc/api/Repositories/CategoryRepository.java
#	src/main/java/com/amc/api/Repositories/Product/ProductRepository.java
#	src/main/java/com/amc/api/Repositories/UserRepository.java
#	src/main/java/com/amc/api/Services/AddressService.java
#	src/main/java/com/amc/api/Services/Product/ProductService.java
#	src/main/java/com/amc/api/Services/UserService.java
#	src/main/java/com/amc/api/Utils/Exceptions.java
#	src/main/resources/application-dev.properties

+++ src/main/java/com/amc/api/Controllers/AddressController.java
@@ -2,7 +2,10 @@ package com.amc.api.Controllers;
 
 
 import com.amc.api.Entities.Address;
+import com.amc.api.Repositories.AddressRepository;
 import com.amc.api.Services.AddressService;
+import com.amc.api.Utils.Exceptions;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -16,32 +19,41 @@ public class AddressController {
     @Autowired
     private AddressService addressService;
 
+    @Autowired
+    private AddressRepository addressRepository;
+
     @GetMapping("/{customerUuid}")
-    public ResponseEntity<Address> getAddressByUuid(@PathVariable(name="customerUuid") String addressUuid) {
+    public ResponseEntity<Address> getAddressByCustomerUuid(@PathVariable(name="customerUuid") String addressUuid) {
+        if (addressRepository.findByUuid(addressUuid) == null)
+            throw new Exceptions.ResourceNotFoundException("Endereço não encontrado.");
         Address address = addressService.findAddressByUuid(addressUuid);
         return address != null ? ResponseEntity.ok().body(address) : ResponseEntity.badRequest().build();
     }
 
     @GetMapping("all/{customerUuid}")
-    public ResponseEntity<Address> getAllAddressesBycustomerUuid(@PathVariable(name="customerUuid") String customerUuid) {
+    public ResponseEntity<Address> getAllAddressesByCustomerUuid(@PathVariable(name="customerUuid") String customerUuid) {
         Address address = addressService.findAllAddressByCustomerUuid(customerUuid);
-        return address != null ? ResponseEntity.ok().body(address) : ResponseEntity.badRequest().build();
+        return address != null ? ResponseEntity.ok().body(address) : ResponseEntity.notFound().build();
     }
 
     @PostMapping("/{customerUuid}")
-    public ResponseEntity<Address> createAddress(@RequestBody List<Address> newAddress, @PathVariable(name="customerUuid") String customerUuid) {
+    public ResponseEntity<Address> createAddress(@Valid @RequestBody Address newAddress, @PathVariable(name="customerUuid") String customerUuid) {
        boolean address = addressService.createAddress(newAddress, customerUuid);
         return address ? ResponseEntity.ok().build() : ResponseEntity.badRequest().build();
     }
 
     @PutMapping("/{addressUuid}")
-    public ResponseEntity<Address> updateAddress(@RequestBody Address updatedAddress, @PathVariable(name="addressUuid") String addressUuid) {
+    public ResponseEntity<Address> updateAddress(@Valid @RequestBody Address updatedAddress, @PathVariable(name="addressUuid") String addressUuid) {
+        if (addressRepository.findByUuid(addressUuid) == null)
+            throw new Exceptions.ResourceNotFoundException("Endereço não encontrado.");
         Address address = addressService.updateAddress(updatedAddress, addressUuid);
         return address != null ? ResponseEntity.ok().body(address) : ResponseEntity.badRequest().build();
     }
 
     @DeleteMapping("/{addressUuid}")
-    public ResponseEntity<Address> deleteAddress(@PathVariable(name="addressUuid") String addressUuid){
+    public ResponseEntity<?> deleteAddress(@PathVariable(name="addressUuid") String addressUuid){
+        if (addressRepository.findByUuid(addressUuid) == null)
+            throw new Exceptions.ResourceNotFoundException("Endereço não encontrado.");
         boolean address = addressService.deleteAddressByUuid(addressUuid);
         return address ? ResponseEntity.noContent().build() : ResponseEntity.badRequest().build();
     }
