+++ src/main/java/com/amc/api/Controllers/CategoryController.java
@@ -4,6 +4,8 @@ package com.amc.api.Controllers;
 import com.amc.api.Entities.Category;
 import com.amc.api.Repositories.CategoryRepository;
 import com.amc.api.Services.CategoryService;
+import com.amc.api.Utils.Exceptions;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -22,8 +24,10 @@ public class CategoryController {
 
     @GetMapping("/{uuid}")
     public ResponseEntity<Category> getCategoryByUuid(@PathVariable("uuid") String categoryUuid ) {
+        if (categoryRepository.findByUuid(categoryUuid) == null)
+            throw new Exceptions.ResourceNotFoundException("Categoria não encontrado.");
         Category category = categoryService.findCategoryByUuid(categoryUuid);
-        return category != null ? ResponseEntity.ok(category) : ResponseEntity.notFound().build();
+        return category != null ? ResponseEntity.ok(category) : ResponseEntity.badRequest().build();
     }
 
     @GetMapping
@@ -33,21 +37,27 @@ public class CategoryController {
     }
 
     @PostMapping
-    public ResponseEntity<Category> createCategory(@RequestBody Category newUser) {
-        Category category = categoryService.createCategory(newUser);
+    public ResponseEntity<Category> createCategory(@Valid @RequestBody Category newCategory) {
+        if (categoryRepository.findBySlug(newCategory.getSlug()) != null || categoryRepository.findByName(newCategory.getName()) != null)
+            throw new Exceptions.DatabaseException("Já existe uma categoria com os dados informados.");
+        Category category = categoryService.createCategory(newCategory);
         return category != null ? ResponseEntity.ok(category) : ResponseEntity.badRequest().build();
     }
 
     @PutMapping("/{uuid}")
-    public ResponseEntity<Category> updateCategory(@PathVariable("uuid") String categoryUuid, @RequestBody Category categoryData) {
+    public ResponseEntity<Category> updateCategory(@Valid @PathVariable("uuid") String categoryUuid, @RequestBody Category categoryData) {
+        if (categoryRepository.findByUuid(categoryUuid) == null)
+            throw new Exceptions.ResourceNotFoundException("Categoria não encontrado.");
         Category category = categoryService.updateCategory(categoryUuid, categoryData);
-        return category != null ? ResponseEntity.ok(category) : ResponseEntity.notFound().build();
+        return category != null ? ResponseEntity.ok(category) : ResponseEntity.badRequest().build();
     }
 
     @DeleteMapping("/{uuid}")
     public ResponseEntity<Category> deleteCategory(@PathVariable("uuid") String categoryUuid ) {
+        if (categoryRepository.findByUuid(categoryUuid) == null)
+            throw new Exceptions.ResourceNotFoundException("Categoria não encontrada.");
         boolean category = categoryService.deleteCategory(categoryUuid);
-        return category ? ResponseEntity.ok().build() : ResponseEntity.notFound().build();
+        return category ? ResponseEntity.ok().build() : ResponseEntity.badRequest().build();
     }
 
 }
